Нагрузочное тестирование. Нагрузка localhost.

+++ database/database/Change-Sets/changelogs-2020-05-05-db-000.user_friends.mysql.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS user_friends (
-  id                INT AUTO_INCREMENT PRIMARY KEY,
-  user_info_id      INT NOT NULL,
-  friend_info_id    INT NOT NULL,
+  id                BIGINT AUTO_INCREMENT PRIMARY KEY,
+  user_info_id      BIGINT NOT NULL,
+  friend_info_id    BIGINT NOT NULL,
   FOREIGN KEY (user_info_id) REFERENCES user_info(id),
   FOREIGN KEY (friend_info_id) REFERENCES user_info(id)
 ) ENGINE = InnoDB;
